Add two profiles: 'dev' and 'test', add GET and POST unit tests for Patient Controller, add separate dataloader for tests, fix controller classes, replace PATCH method in DoctorController with PUT and fix the updating method in UserServiceImpl class

+++ src/main/java/com/bairei/restspringboot/services/UserServiceImpl.java
@@ -2,9 +2,14 @@ package com.bairei.restspringboot.services;
 
 import com.bairei.restspringboot.domain.Role;
 import com.bairei.restspringboot.domain.User;
+import com.bairei.restspringboot.exceptions.InternalServerException;
+import com.bairei.restspringboot.exceptions.UserNotFoundException;
 import com.bairei.restspringboot.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -42,6 +47,27 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
+    public User put(User user) throws UserNotFoundException, InternalServerException, Exception {
+
+        if (user.getId() == null) return this.save(user);
+
+        User detachedUser = userRepository.getOne(user.getId());
+        if(detachedUser == null) throw new UserNotFoundException(user.getId());
+        if(user.getEmail() != null) detachedUser.setEmail(user.getEmail());
+        if(user.getName() != null) detachedUser.setName(user.getName());
+        if(user.getSurname() != null) detachedUser.setSurname(user.getSurname());
+        if(!user.getRoles().equals(detachedUser.getRoles())) detachedUser.setRoles(user.getRoles());
+        if(user.getPassword() != null && user.getConfirmPassword() != null && user.getPassword().equals(user.getConfirmPassword())){
+            detachedUser.setPassword(passwordEncoder.encode(user.getPassword()));
+            detachedUser.setConfirmPassword(passwordEncoder.encode(user.getConfirmPassword()));
+        }
+        if (user.getPassword() != null && !user.getPassword().equals(user.getConfirmPassword())) {
+            throw new InternalServerException(new Exception("Please provide correct password and / or confirmation"));
+        }
+        return userRepository.save(detachedUser);
+    }
+
+    @Override
     public List<User> findAll() {
         return userRepository.findAll();
     }
@@ -58,7 +84,10 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Integer id) throws Exception, UserNotFoundException {
+        if(this.findOne(id) == null) throw new UserNotFoundException(id);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (auth.getName().equalsIgnoreCase(this.findOne(id).getEmail())) throw new Exception("You can't remove your own account!");
         userRepository.deleteById(id);
     }
